63 - MealRepo & AuthRepo & ProfileRepo fixes (#70)
* Creating repos

* repos fixes

* auth and meal repos fixes

* fixes

* fixes

* little fixes

+++ app/src/main/java/dev/calorai/mobile/features/meal/data/api/MealApi.kt
@@ -11,9 +11,10 @@ import retrofit2.http.Query
 
 
 interface MealApi {
+
     @GET("daily-meal/{userId}")
     suspend fun getDailyMeal(
-        @Path("userId") userId: Int,
+        @Path("userId") userId: Long,
         @Query("date") date: String? = null
     ): Response<GetDailyMealResponse>
